From ea793cb9f744cd2942b66385118161b880de958a Mon Sep 17 00:00:00 2001 From: Luiz Augusto von Dentz <luiz.von.dentz@intel.com> Date: Tue, 19 Jun 2012 12:44:43 +0300 Subject: [PATCH] input: Fix not adding watches when io channel is connected
This can leave dangling pointers in case one of the channel is never
connected which cause -EALREADY to be returned by
input_device_set_channel next time the device attempts to connect.

For the same reason the code path when acting as client now add the
watch as soon as the connection completes instead when both channels
are connected.