From: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com> Subject: Allow more space for struct hci_conn_info
HACK ALERT: This is very brittle. It will likely blow up in flames if the
hci_conn_info / hci_conn_info_req structures change.

Deal with the alignment requirements for ARM by requesting more memory from
malloc, double the size of 'struct hci_conn_info'.

This is required because as you malloc the memory space required to hold the
request and response for the HCIGETCONNINFO ioctl, it doesn't take into
consideration that on some architectures, the kernel may be padding data for
you as required by the architecture. Doing so means we get more data for the
pointer than we allocated, and freeing that data will therefore fail:

 *** Error in `./toto': free(): invalid next size (fast): 0x0009f008 ***

With this patch we're simply requesting more space (twice the size of
struct hci_conn_info; so 32 bytes); which means we have 16 more bytes
allocated for us in hope to catch the extra padding from the kernel.

I've calculated the extra to about 12 bytes if all variables are aligned
to a word boundary in that struct.