Author: Ricardo Salveti de Araujo <ricardo.salveti@canonical.com> Description: Don't overwrite default discoverable timeout
This is a backport of the upstream patch 2ea895e, which disables
the values used as example in main.conf, as the code itself already
provides a default value for them.

As a side effect, the default discoverable timeout will now be 3,
instead of forever.

Bug-Ubuntu: https://bugs.launchpad.net/ubuntu/+source/bluez/+bug/1419874